--- a/.zenodo.json
+++ b/.zenodo.json
@@ -36,11 +36,16 @@
          "affiliation": "Université de Montréal",
          "name": "Banville, Francis",
          "orcid": "0000-0001-9051-0597"
-      }, 
+      },
       {
       "affiliation": "School of Mathematics and Statistics, University of Canterbury",
       "name": "Dalla Riva, Giulio V.",
       "orcid": "0000-0002-3454-0633"
+      },
+      {
+      "affiliation": "Tokyo University of Technology",
+      "name": "Shuta Ito",
+      "orcid": "0000-0003-0030-4648"
       }
    ],
    "access_right": "open",