--- a/src/modularity/utilities.jl
+++ b/src/modularity/utilities.jl
@@ -33,23 +33,25 @@ function Q(N::T, L::Dict{E,Int64}) where {T<:AbstractEcologicalNetwork,E}
 
     # Degrees
     dkin, dkout = degree_in(N), degree_out(N)
-    kin = map(x -> dkin[x], species(N; dims=2))
-    kout = map(x -> dkout[x], species(N; dims=1))
+    names1 = species(N; dims=2)
+    names2 = species(N; dims=1)
+    kin = map(x -> dkin[x], names1)
+    kout = map(x -> dkout[x], names2)
 
     # Value of m -- sum of weights, total number of int, ...
     m = links(N)
+    modularity = 0.0
+    for c in 1:length(kin)
+      for r in 1:length(kout)
+        name1 = names1[c]
+        name2 = names2[r]
+        if L[name1] != L[name2] continue end
+        modularity += (N.edges[r, c]) - (kout[r] * kin[c] / m)
+      end
+    end
+    modularity /= m
 
-    # Null model
-    kikj = (kout .* kin')
-    Pij = kikj ./ m
-
-    # Difference
-    diff = N.edges .- Pij
-
-    # Diff × delta
-    dd = diff .* δ(N,L)
-
-    return sum(dd)/m
+    return modularity
 end
 
 """