--- a/src/modularity/utilities.jl
+++ b/src/modularity/utilities.jl
@@ -41,12 +41,12 @@ function Q(N::T, L::Dict{E,Int64}) where {T<:AbstractEcologicalNetwork,E}
     # Value of m -- sum of weights, total number of int, ...
     m = links(N)
     modularity = 0.0
-    for c in 1:length(kin)
-      for r in 1:length(kout)
-        name1 = names1[c]
-        name2 = names2[r]
+    for i in 1:length(kin)
+      for j in 1:length(kout)
+        name1 = names1[i]
+        name2 = names2[j]
         if L[name1] != L[name2] continue end
-        modularity += (N.edges[r, c]) - (kout[r] * kin[c] / m)
+        modularity += (N.edges[j, i]) - (kout[j] * kin[i] / m)
       end
     end
     modularity /= m