@tpoisot 2021-03-05
Really nice - I'm going to tag a new release with this, it's especially efficient now that the networks are sparse matrices!